+++ diesel_tests/tests/compile-fail/select_sql_still_ensures_result_type.rs
@@ -13,6 +13,6 @@ table! {
 fn main() {
     let connection = Connection::establish("").unwrap();
     let select_count = users::table.select_sql::<types::BigInt>("COUNT(*)");
-    let count = connection.query_one::<_, String>(select_count).unwrap();
+    let count = select_count.get_result::<String>(&connection).unwrap();
     //~^ ERROR E0277
 }